Removed spring dependency
Spring is no longer a dependency. However the the usage has changed a
little. Now interface is more like ForkJoinPool or other Executers in
Java Standard library. You create instance of AsyncRequestSerializer
passing the config. And then call submit method on that instance with
with (key, request) and get the Future back.